debug clear and add check password with each command

+++ Library/src/main/java/ru/nanikon/FlatCollection/commands/RemoveAnyByTransportCommand.java
@@ -25,6 +25,9 @@ public class RemoveAnyByTransportCommand implements Command, Serializable {
      */
     @Override
     public String execute(DBManager manager) {
+        if (!manager.chekUser(login, password)) {
+            return "Ой, вы там в приложении что-то напортачили и мы то ли логин не найдем, то ли пароль для него не тот. Перезайдите нормально!";
+        }
         Transport transport = ((TransportArg) params[0]).getValue();
         return manager.deleteByTransport(transport.name(), login);
     }
