+++ Library/src/main/java/ru/nanikon/FlatCollection/commands/RemoveCommand.java
@@ -24,6 +24,9 @@ public class RemoveCommand implements Command, Serializable {
      */
     @Override
     public String execute(DBManager manager) {
+        if (!manager.chekUser(login, password)) {
+            return "Ой, вы там в приложении что-то напортачили и мы то ли логин не найдем, то ли пароль для него не тот. Перезайдите нормально!";
+        }
         int id = ((IntArg) params[0]).getValue();
         return manager.deleteById(id, login);
     }
